opencsd: bugfix - ETMv4, PTM - output packet error.
Initilisation issue was causing a crash in the debug build tests in windows.
Generic elements not initialising some parts of the structure - enhanced
printing in 0v003 saw out of range in array. Assume linux passed due to
zero initialisation of these elements.

Further invesigation showed that the .isa element on the generic output
packet not always set in PTM and ETMv4

Signed-off-by: Mike Leach <mike.leach@linaro.org>